requirements.txt
freetype_py==2.3.0
numpy==1.24.2
path.py==12.5.0
Pillow==9.4.0
pygame==2.1.3
PyOpenGL==3.1.6
